[tilegx] Add various interfaces for TileGX.
  * try to fix those compile errors exposed by "make native"
    as much as possible

  * target-independent file "src/frame.h" modified, because TileGX
    has 64 registers, the original 32bit bitmap designed for
    x86/mips/arm is not enough.

+++ src/frames.h
@@ -35,7 +35,7 @@
 namespace v8 {
 namespace internal {
 
-typedef uint32_t RegList;
+typedef uint64_t RegList;
 
 // Get the number of registers in a given register list.
 int NumRegs(RegList list);
